Makes axis text labels flush on the ends

A covenience function intended for basic, fixed-scale plots only (i.e. does not handle free scales in facets).

You need to pass in a ggplot2 object to this function. It can't be +'d in a chain of geoms, coords, scales, themes, etc. It also builds the plot (but does not display it) so if the plt takes a while (i.e. has lots of data or transforms) this will also take a while.

easy_flush_ticks(gg, flush = "XY", teach = TRUE, plot = FALSE)

Arguments

gg
ggplot2 plot object

flush
either "X" or "Y" or "XY" to flush individual or both axes. Default: both.

teach
if TRUE then display theme() statements and copy them to the clipboard

plot
if FALSE then the ggplot object will be returned invisibly

Value

ggplot2 object with theme() elements added

Details

If clipr is installed and cat is TRUE then the theme(...) statements will be copied to the clipboard as well as displayed (via cat()) in the console.

Note

Intended for basic, fixed-scale plots only (i.e. does not handle free scales in facets). Also, the clipboard operations will only occur if clipr is installed.